// src/client/common/types.ts
export interface IPythonSettings {
    readonly pythonPath: string;
    readonly venvPath: string;
    readonly venvFolders: string[];
    readonly jediEnabled: boolean;
    readonly jediPath: string;
    readonly jediMemoryLimit: number;
    readonly devOptions: string[];
    readonly linting?: ILintingSettings;
    readonly formatting?: IFormattingSettings;
    readonly unitTest?: IUnitTestSettings;
    readonly autoComplete?: IAutoCompleteSettings;
    readonly terminal: ITerminalSettings;
    readonly sortImports?: ISortImportSettings;
    readonly workspaceSymbols?: IWorkspaceSymbolSettings;
    readonly envFile: string;
    readonly disablePromptForFeatures: string[];
    readonly disableInstallationChecks: boolean;
    readonly globalModuleInstallation: boolean;
}
export interface ISortImportSettings {
    readonly path: string;
    readonly args: string[];
}

export interface IUnitTestSettings {
    readonly promptToConfigure: boolean;
    readonly debugPort: number;
    readonly debugHost?: string;
    readonly nosetestsEnabled: boolean;
    nosetestPath: string;
    nosetestArgs: string[];
    readonly pyTestEnabled: boolean;
    pyTestPath: string;
    pyTestArgs: string[];
    readonly unittestEnabled: boolean;
    unittestArgs: string[];
    cwd?: string;
    readonly useExperimentalDebugger?: boolean;
}
// ...
export interface ILintingSettings {
    readonly enabled: boolean;
    readonly ignorePatterns: string[];
    readonly prospectorEnabled: boolean;
    readonly prospectorArgs: string[];
    readonly pylintEnabled: boolean;
    readonly pylintArgs: string[];
    readonly pep8Enabled: boolean;
    readonly pep8Args: string[];
    readonly pylamaEnabled: boolean;
    readonly pylamaArgs: string[];
    readonly flake8Enabled: boolean;
    readonly flake8Args: string[];
    readonly pydocstyleEnabled: boolean;
    readonly pydocstyleArgs: string[];
    readonly lintOnSave: boolean;
    readonly maxNumberOfProblems: number;
    readonly pylintCategorySeverity: IPylintCategorySeverity;
    readonly pep8CategorySeverity: IPep8CategorySeverity;
    readonly flake8CategorySeverity: Flake8CategorySeverity;
    readonly mypyCategorySeverity: IMypyCategorySeverity;
    prospectorPath: string;
    pylintPath: string;
    pep8Path: string;
    pylamaPath: string;
    flake8Path: string;
    pydocstylePath: string;
    mypyEnabled: boolean;
    mypyArgs: string[];
    mypyPath: string;
    readonly pylintUseMinimalCheckers: boolean;
}
export interface IFormattingSettings {
    readonly provider: string;
    autopep8Path: string;
    readonly autopep8Args: string[];
    yapfPath: string;
    readonly yapfArgs: string[];
}
export interface IAutoCompleteSettings {
    readonly addBrackets: boolean;
    readonly extraPaths: string[];
    readonly preloadModules: string[];
    readonly showAdvancedMembers: boolean;
}
export interface IWorkspaceSymbolSettings {
    readonly enabled: boolean;
    tagFilePath: string;
    readonly rebuildOnStart: boolean;
    readonly rebuildOnFileSave: boolean;
    readonly ctagsPath: string;
    readonly exclusionPatterns: string[];
}
export interface ITerminalSettings {
    readonly executeInFileDir: boolean;
    readonly launchArgs: string[];
    readonly activateEnvironment: boolean;
}
export interface IPythonAnalysisEngineSettings {
    readonly showAdvancedMembers: boolean;
}

export const IConfigurationService = Symbol('IConfigurationService');

export interface IConfigurationService {
    getSettings(resource?: Uri): IPythonSettings;
    isTestExecution(): boolean;
    updateSettingAsync(setting: string, value?: {}, resource?: Uri, configTarget?: ConfigurationTarget): Promise<void>;
}